+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Added
 
 - [fathom-rest] Allow declaration of @Accepts separately from @Produces
+- [fathom-rest] Add RouteInterceptor concept for controller routes
 
 #### Removed